+++ forst/nodert/host.go
@@ -149,6 +149,9 @@ func spawnHostProcess(cmd HostSpawnCommand, workDir string, log *logrus.Logger) 
 	}
 	execCmd.Env = cmd.Env
 	execCmd.SysProcAttr = hostSessionAttrs()
+	if SupportsInvokeAuthFDHandoff() && len(cmd.ExtraFiles) > 0 {
+		execCmd.ExtraFiles = append([]*os.File(nil), cmd.ExtraFiles...)
+	}
 
 	stdout, err := execCmd.StdoutPipe()
 	if err != nil {